Remove duplicate utilities.
